requirements.txt
python-telegram-bot==20.7
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0